# src/main/resources/swagger-ui/swagger.yaml
openapi: 3.0.3

info:
  title: Users API
  description: API for managing system users.
  version: 1.0.0

tags:
  - name: Users
    description: Operations related to users.

paths:

  /users:
    get:
      summary: Search and filter users with pagination
      operationId: listUsers
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/NameQuery'
        - $ref: '#/components/parameters/EmailQuery'
        - $ref: '#/components/parameters/UsernameQuery'
        - $ref: '#/components/parameters/ActiveQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Users list returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUsersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create a new user
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User successfully created.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  /users/{id}:
    get:
      summary: Get user by ID
      operationId: getUserById
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: User found successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      summary: Update an existing user
      operationId: updateUser
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '204':
          description: User successfully updated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete a user
      operationId: deleteUser
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: User successfully deleted.
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /users/username/{username}:
    get:
      summary: Get user by username
      operationId: getUserByUsername
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/PathUsername'
      responses:
        '200':
          description: User found successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

components:

  schemas:
    UserRequest:
      type: object
      required:
        - username
        - name
        - email
        - activeUser
      properties:
        username:
          type: string
          example: jdoe
        name:
          type: string
          example: John Doe
        password:
          type: string
          writeOnly: true
          example: Password@123
        email:
          type: string
          format: email
          example: john.doe@email.com
        activeUser:
          type: boolean
          example: true

    UserResponse:
      allOf:
        - type: object
          properties:
            id:
              type: string
              readOnly: true
              example: 652ff3a9b1c2d40012ab45de
        - $ref: '#/components/schemas/UserRequest'
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
              readOnly: true
              example: 2025-10-11T10:15:30Z
            updatedAt:
              type: string
              format: date-time
              readOnly: true
              example: 2025-10-11T10:20:45Z

    PagedUsersResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        totalElements:
          type: integer
          format: int64
          example: 120
        limit:
          type: integer
          example: 10
        offset:
          type: integer
          example: 0

    ApiError:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid parameters.
        path:
          type: string
          example: /users
        timestamp:
          type: string
          format: date-time

  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: User ID

    PathUsername:
      name: username
      in: path
      required: true
      schema:
        type: string
      description: User username

    UsernameQuery:
      name: username
      in: query
      schema:
        type: string
      description: Filter users by username

    ActiveQuery:
      name: active
      in: query
      schema:
        type: boolean
      description: Filter users by active status

    NameQuery:
      name: name
      in: query
      schema:
        type: string
      description: Filter users by name

    EmailQuery:
      name: email
      in: query
      schema:
        type: string
        format: email
      description: Filter users by email

    Limit:
      name: limit
      in: query
      schema:
        type: integer
        example: 10
        minimum: 1
      description: Number of records to return

    Offset:
      name: offset
      in: query
      schema:
        type: integer
        example: 0
        minimum: 0
      description: Number of records to skip

  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    Unauthorized:
      description: Authentication failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    Forbidden:
      description: Access denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'